The RAM cache is now working on local portlets that are dependent on the context, i.e. entire slots can be cached in RAM. Cache parameters are defined in the 'common portlet' schema as a list ('url', 'i18n', 'user' ... ). 'url' means that the portlet is url-dependent, 'i18n' means that the portlet is language-dependent, 'user' means that that portlet is dependent on the user, ...
This should ideally be set in the FTI since these values are defined per portlet type and they never change.
Instead they are now defined in CPSPortlets/skins/cpsportlets_default/getCPSPortletCacheParams.py.

The cache parameters are then passed to CPSSkins with getCacheParams() and CPSSkins computes a cache index. The portlet can also pass a custom cache index with getCustomCacheIndex() for index values that CPSSkins does not need to know any details about.

In terms of performance, for instance the 'Content creation' portlet is now displayed in 0.05 seconds instead of 0.7 seconds.