Implemented a local RAM cache manager for local portlets - the cache manager is located in the Portlets Tool.
Performance and responsiveness are highly improved since the cache is now invalidated on a portlet level and not on a slot level. Hence portlets can be moved arround and still reside in the cache. There are also much fewer entries in the cache index.

There is no lifetime associated to a portlet since the invalidation will be entirely managed through the event service.

The cache management for global portlets is still done by CPSSkins.