compatibility with CMF/Plone
CMF (and Plone) require that an ID be provided before creating a document.

We do not want to handle the ID check part so the "Add Item" portlet redirects the user to a simplified folder factory form instead with an ID input field - the document type is already selected. see: http://www.medic.chalmers.se/~jmo/CPS/additem-cmf.png