Implemented event-triggered invalidation of the portlet cache.
The 'content portlet' (that does catalog searches) get invalidated on the following events: workflow_publish, workflow_modify, sys_del_object.

This must be fine-tuned and narrowed down to event objects in the case of 'workflow_modify' and 'sys_del_object'.

for the 'Content portlet' the improvement in rendering times is from 0.9s to 0.1s.