Marker interface for Text Portlets
Necessary so that custom rendering modes can be used.
The result is actually the z3-style equivalent of Custom Portlets, and
somewhat more powerful, since it also provides a user-editable text that can
be used together with whatever a Portlets view class can handle.

Custom Portlets should probably be deprecated, if it mattered.